Add a known_hosts module
This is an ansible module for managing the ssh known_hosts file. It's
useful because known_hosts files often contain hashed entries, which
means management by regular templates is troublesome - so this module
uses ssh-keygen to search for entries that match a particular host,
and to remove old entries if updating.